+++ package.json
@@ -2,7 +2,6 @@
   "name": "pushup-site",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:4000",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
@@ -12,6 +11,7 @@
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "firebase": "^9.8.2",
+    "node-schedule": "^2.1.0",
     "react": "^18.1.0",
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.1.0",
