+++ src/App/Store.js
@@ -5,13 +5,11 @@ import { composeWithDevTools } from "redux-devtools-extension";
 import userReducer from "./Redux/User";
 import authReducer from "./Redux/Auth";
 import leaderboardReducer from "./Redux/Leaderboard";
-import pollingReducer from "./Redux/Polling";
 
 const reducer = combineReducers({
   user: userReducer,
   auth: authReducer,
-  leaderboard: leaderboardReducer,
-  polling: pollingReducer
+  leaderboard: leaderboardReducer
 });
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({ collapsed: true }))
